+++ client/package.json
@@ -11,7 +11,23 @@
   },
   "dependencies": {
     "react": "16.9.0",
-    "react-native": "0.61.5"
+    "react-native": "0.61.5",
+    "@apollo/react-components": "^3.1.3",
+    "@react-native-community/async-storage": "^1.8.1",
+    "@react-native-community/masked-view": "^0.1.7",
+    "@react-navigation/bottom-tabs": "^5.2.0",
+    "@react-navigation/native": "^5.0.10",
+    "@react-navigation/stack": "^5.2.0",
+    "apollo-boost": "^0.4.7",
+    "graphql": "^14.6.0",
+    "moment": "^2.24.0",
+    "react-native-gesture-handler": "^1.6.0",
+    "react-native-onboarding-swiper": "^1.1.0",
+    "react-native-range-datepicker": "^1.9.1",
+    "react-native-reanimated": "^1.7.0",
+    "react-native-safe-area-context": "^0.7.3",
+    "react-native-screens": "^2.3.0",
+    "react-native-vector-icons": "^6.6.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.7",
